\name{plot.lasso}
\alias{plot.lasso}
\title{
Plot Function for "lasso"
}
\description{
Visualize the solution path of regression estimate corresponding to regularization paramters.
}
\usage{
\method{plot}{lasso}(x, ...)
}
\arguments{
  \item{x}{
  An object with S3 class \code{"lasso"}.
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}

\author{
Xingguo Li, Tuo Zhao, Tong Zhang and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}

\seealso{
\code{\link{picasso}} and \code{\link{picasso-package}}.
}
