\name{plot.logit}
\alias{plot.logit}
\title{
Plot Method for an Object with S3 Class \code{"logit"}
}
\description{
Plot coefficient paths across regularization parameters.
}
\usage{
\method{plot}{logit}(x, ...)
}
\arguments{
  \item{x}{
  An object with S3 class \code{"logit"}.
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}
\value{
No return value, called for side effects (drawing the coefficient-path plot).
}

\author{
Jason Ge, Xingguo Li, Haoming Jiang, Mengdi Wang, Tong Zhang, Han Liu and Tuo Zhao\cr
Maintainer: Tuo Zhao <tourzhao@gatech.edu>
}

\seealso{
\code{\link{picasso}} and \code{\link{picasso-package}}.
}
