\name{plot.scio}
\alias{plot.scio}
\title{
Plot Function for "scio"
}
\description{
Plot sparsity level information and 3 typical sparse graphs from the graph path.
}
\usage{
\method{plot}{scio}(x, align = FALSE, \dots)
}
\arguments{
  \item{x}{
  An object with S3 class \code{"scio"} 
  }
  \item{align}{
  If \code{align = FALSE}, 3 plotted graphs are aligned
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}

\author{
Xingguo Li, Tuo Zhao, Tong Zhang and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
\seealso{
\code{\link{picasso}} and \code{\link{picasso-package}}
}
