\name{plot.sim}
\alias{plot.sim}
\title{
Plot Function for "sim"
}
\description{
Visualize the covariance matrix, the empirical covariance matrix, the adjacency matrix and the graph pattern of the true graph structure.
}
\usage{
\method{plot}{sim}(x, ...)
}
\arguments{
  \item{x}{
  An object with S3 class \code{"sim"} 
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}

\author{
Xingguo Li, Tuo Zhao, Tong Zhang and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
\seealso{
\code{\link{scio.generator}}, \code{\link{picasso}} and \code{\link{picasso-package}}
}
