\name{print.mvr}
\alias{print.mvr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print Function for an object with S3 class \code{"mvr"}
}
\description{
Print a summary of the information about an object with S3 class \code{"mvr"}.
}
\usage{
\method{print}{mvr}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
  An object with S3 class \code{"mvr"}.
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}
\details{
This call simply outlines the options used for computing a mvr object.
}
%\value{
%}
%\references{
%}

\author{
Xingguo Li, Tuo Zhao and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
%\note{
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{picasso.mvr}} and \code{\link{picasso-package}}.
}
