\name{print.sim}
\alias{print.sim}
\title{
Print Function for for an object with S3 class \code{"sim"}
}
\description{
Print the information about the sample size, the dimension, the pattern and sparsity of the true graph structure.
}
\usage{
\method{print}{sim}(x, ...)
}
\arguments{
  \item{x}{
  An object with S3 class \code{"sim"}.
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}
\author{
Xingguo Li, Tuo Zhao, Tong Zhang and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
\seealso{
\code{\link{picasso}} and \code{\link{scio.generator}}
}
