\name{picasso-package}
\alias{picasso-package}
\docType{package}
\title{
picasso: Pathwise Calibrated Sparse Shooting Algorithm
}
\description{
Provides the implementation of a new family of efficient algorithms, called PathwIse CalibrAted Sparse Shooting AlgOrithm (picasso), for a variety of sparse learning problems, including Sparse Linear Regression, Sparse Logistic Regression, Sparse Column Inverse Operator and Sparse Multivariate Regression. Different types of active set identification schemes are implemented, such as cyclic search, greedy search, stochastic search and proximal gradient search. Besides, the package provides the choices between convex (L1 norm) and non-convex (MCP and SCAD) regularizations. Moreover, group regularization, such as group Lasso, group MCP and group SCAD, are also implemented for Sparse Linear Regression, Sparse Logistic Regression and Sparse Multivariate Regression.
}
\details{
\tabular{ll}{
Package: \tab picasso\cr
Type: \tab Package\cr
Version: \tab 0.3.0\cr
Date: \tab 2014-10-18\cr
License: \tab GPL-2\cr
}
}

\author{
Xingguo Li, Tuo Zhao, and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}

\seealso{
\code{\link{picasso}} and \code{\link{scio.generator}}.
}
