\name{picasso-package}
\alias{picasso-package}
\docType{package}
\title{
PICASSO: Pathwise Calibrated Sparse Shooting Algorithm
}
\description{
Computationally efficient tools for fitting sparse generalized linear models
with convex and non-convex penalties (L1, MCP, and SCAD). The package uses
pathwise warm starts, active-set updates, and screening rules for fast model
training along a regularization path.
}
\details{
\tabular{ll}{
Package: \tab picasso\cr
Type: \tab Package\cr
Version: \tab 1.4.1\cr
Date: \tab 2026-02-18\cr
License: \tab GPL-3\cr
}
}

\author{
Jason Ge, Xingguo Li, Mengdi Wang, Tong Zhang, Han Liu and Tuo Zhao\cr
Maintainer: Tuo Zhao <tourzhao@gatech.edu>
}

\seealso{
\code{\link{picasso}}.
}
