\name{plot.gaussian}
\alias{plot.gaussian}
\title{
Plot Function for "gaussian"
}
\description{
Visualize the solution path of regression estimate corresponding to regularization paramters.
}
\usage{
\method{plot}{gaussian}(x, ...)
}
\arguments{
  \item{x}{
  An object with S3 class \code{"gaussian"}.
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}

\author{
Jason Ge, Xingguo Li, Mengdi Wang, Tong Zhang, Han Liu and Tuo Zhao\cr
Maintainer: Jason Ge <jiange@princeton.edu>
}

\seealso{
\code{\link{picasso}} and \code{\link{picasso-package}}.
}
