\name{plot.logit}
\alias{plot.logit}
\title{
Plot Function for "logit"
}
\description{
Visualize the solution path of regression estimates across regularization parameters.
}
\usage{
\method{plot}{logit}(x, ...)
}
\arguments{
  \item{x}{
  An object with S3 class \code{"logit"}.
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}

\author{
Jason Ge, Xingguo Li, Mengdi Wang, Tong Zhang, Han Liu and Tuo Zhao\cr
Maintainer: Tuo Zhao <tourzhao@gatech.edu>
}

\seealso{
\code{\link{picasso}} and \code{\link{picasso-package}}.
}
