\name{print.logit}
\alias{print.logit}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print Function for an object with S3 class \code{"logit"}
}
\description{
Print a summary of the information about an object with S3 class \code{"logit"}.
}
\usage{
\method{print}{logit}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
  An object with S3 class \code{"logit"}.
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}
\details{
This call simply outlines the options used for computing a logit object.
}
%\value{
%}
%\references{
%}

\author{
Jason Ge, Xingguo Li, Mengdi Wang, Tong Zhang, Han Liu and Tuo Zhao\cr
Maintainer: Tuo Zhao <tourzhao@gatech.edu>
}
%\note{
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{picasso}} and \code{\link{picasso-package}}.
}
