\name{print.sqrtlasso}
\alias{print.sqrtlasso}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print Function for an object with S3 class \code{"sqrtlasso"}
}
\description{
Print a summary of the information about an object with S3 class \code{"sqrtlasso"}.
}
\usage{
\method{print}{sqrtlasso}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
  An object with S3 class \code{"sqrtlasso"}.
  }
  \item{\dots}{
  Arguments to be passed to methods.
  }
}
\details{
This call simply outlines the options used for computing a lasso object.
}
%\value{
%}
%\references{
%}

\author{
Jason Ge, Xingguo Li, Mengdi Wang, Tong Zhang, Han Liu and Tuo Zhao\cr
Maintainer: Jason Ge <jiange@princeton.edu>
}
%\note{
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{picasso}} and \code{\link{picasso-package}}.
}
